virt-api: Extract lifecycle handlers in their own file
